+++ common/src/main/java/com/faboslav/variantsandventures/common/entity/mob/VerdantEntity.java
@@ -1,15 +1,21 @@
 package com.faboslav.variantsandventures.common.entity.mob;
 
 import com.faboslav.variantsandventures.common.init.VariantsAndVenturesSoundEvents;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundEvent;
+import net.minecraft.sounds.SoundEvents;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.monster.AbstractSkeleton;
 import net.minecraft.world.entity.projectile.AbstractArrow;
 import net.minecraft.world.entity.projectile.Arrow;
+import net.minecraft.world.entity.projectile.Projectile;
+import net.minecraft.world.entity.projectile.ProjectileUtil;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.level.Level;
 import org.jetbrains.annotations.Nullable;
 
@@ -44,6 +50,23 @@ public final class VerdantEntity extends AbstractSkeleton
 
 	public SoundEvent getStepSound() {
 		return VariantsAndVenturesSoundEvents.ENTITY_VERDANT_STEP.get();
+	}
+
+	@Override
+	public void performRangedAttack(LivingEntity target, float velocity) {
+		ItemStack itemStack = this.getItemInHand(ProjectileUtil.getWeaponHoldingHand(this, Items.BOW));
+		ItemStack itemStack2 = this.getProjectile(itemStack);
+		AbstractArrow abstractArrow = this.getArrow(itemStack2, velocity, itemStack);
+		double d = target.getX() - this.getX();
+		double e = target.getY(0.3333333333333333) - abstractArrow.getY();
+		double f = target.getZ() - this.getZ();
+		double g = Math.sqrt(d * d + f * f);
+		Level var15 = this.level();
+		if (var15 instanceof ServerLevel serverLevel) {
+			Projectile.spawnProjectileUsingShoot(abstractArrow, serverLevel, itemStack2, d, e + g * 0.20000000298023224, f, 1.6F, (float)(14 - serverLevel.getDifficulty().getId() * 4));
+		}
+
+		this.playSound(VariantsAndVenturesSoundEvents.ENTITY_VERDANT_ATTACK.get(), 1.0F, 1.0F / (this.getRandom().nextFloat() * 0.4F + 0.8F));
 	}
 
 	@Override